requirements.txt
streamlit>=1.30.0
langgraph>=0.0.15
langchain>=0.1.0
langchain-google-genai>=0.0.6
openai-whisper>=20231117
TTS>=0.22.0  # Coqui TTS
google-cloud-texttospeech>=2.14.1
python-dotenv>=1.0.0
requests>=2.31.0
